Toggle email syncing for existing accounts (#46)
* Allow to selectively toggle email sync from accounts API

* Switch accounts to "initial" state if they weren't synced for a while

When an account is paused and resumed at a later stage, we want to put folders
back into "initial" state. This will ensure that the latest messages are synced
first, rather than oldest first. Note that this won't trigger a complete resync
of the account, it will just sync missing messages.