Explicit `return None` + optimize the case of self-sending.
@alecrosenbaum said:
> Personally I prefer an explicit `return None` because it's less likely that
> I'll miss it when reading through a function.

I cannot personally relate to it, but if there's anyone who benefits from
`return None`, then I'm happy to oblige.

While making this change, I also noticed that one of the `return` statements
within the thread-match loop only looked at the `message` param and didn't
need any additional data (also, none of the `return` statements above it
could've matched if the `message` didn't have a `from_addr` and a `to_addr`).
Because of that, we can move that check all the way up and hence, in case of
self-sent emails, we don't even need to fetch anything from the database.