Use UTF-8 for Content-Disposition header
As of Werkzeug 2.3.0:

"Passing bytes where strings are expected is deprecated, as well as the
charset and errors parameters in many places. Anywhere that was
annotated, documented, or tested to accept bytes shows a warning.
Removing this artifact of the transition from Python 2 to 3 removes a
significant amount of overhead in instance checks and encoding cycles.
In general, always work with UTF-8, the modern HTML, URL, and HTTP
standards all strongly recommend this. #2602"

https://werkzeug.palletsprojects.com/en/3.0.x/changes/#version-2-3-0