Update werkzeug to 3.0.1 (#549)
* Update werkzeug to 3.0.1

* Update to match new get_json() behavior

https://werkzeug.palletsprojects.com/en/3.0.x/wrappers/#werkzeug.wrappers.Request.get_json
"Changed in version 2.3: Raise a 415 error instead of 400."

* Use UTF-8 for Content-Disposition header

As of Werkzeug 2.3.0:

"Passing bytes where strings are expected is deprecated, as well as the
charset and errors parameters in many places. Anywhere that was
annotated, documented, or tested to accept bytes shows a warning.
Removing this artifact of the transition from Python 2 to 3 removes a
significant amount of overhead in instance checks and encoding cycles.
In general, always work with UTF-8, the modern HTML, URL, and HTTP
standards all strongly recommend this. #2602"

https://werkzeug.palletsprojects.com/en/3.0.x/changes/#version-2-3-0

* Use send_file to create file download response (#555)

* Use send_file

* Fix hypothesis int ranges

---------

Co-authored-by: Paweł Piotr Przeradowski <showerproof86@gmail.com>